requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
spacy>=3.7.0
networkx>=3.1
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
tqdm>=4.65.0
rouge-score>=0.1.2
nltk>=3.8.1
rapidfuzz>=3.5.0
pyvis>=0.3.2
scipy>=1.11.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Optional: for transformer-based spaCy
# spacy-transformers>=1.2.0

# Dev / notebook
ipykernel>=6.25.0
jupyter>=1.0.0
